Modularized `bee.generative` & feature implementation
