+++ bee.fetched/src/processor/kotlin/com/beeproduced/bee/fetched/codegen/BeeFetchedOption.kt
@@ -1,0 +1,15 @@
+package com.beeproduced.bee.fetched.codegen
+
+import com.beeproduced.bee.fetched.annotations.BeeFetched
+
+/**
+ *
+ *
+ * @author Kacper Urbaniec
+ * @version 2023-10-14
+ */
+object BeeFetchedOption {
+    const val scanPackage = "fetchedScanPackage"
+    const val packageName = "fetchedPackageName"
+    val beeFetchedAnnotationName = BeeFetched::class.qualifiedName!!
+}
