+++ bee.generative/src/main/kotlin/com/beeproduced/bee/generative/processor/DataProcessor.kt
@@ -1,7 +1,9 @@
 package com.beeproduced.bee.generative.processor
 
+import com.beeproduced.bee.generative.BeeGenerativeConfig
 import com.beeproduced.bee.generative.BeeGenerativeFeature
 import com.beeproduced.bee.generative.BeeGenerativeInput
+import com.beeproduced.bee.generative.BeeGenerativeSymbols
 import com.google.devtools.ksp.KspExperimental
 import com.google.devtools.ksp.processing.*
 import com.google.devtools.ksp.symbol.KSAnnotated
@@ -14,148 +16,81 @@ import java.util.*
  * @author Kacper Urbaniec
  * @version 2023-08-07
  */
+typealias Options = Map<String, String>
+
 class DataProcessor(
     private val codeGenerator: CodeGenerator,
     private val logger: KSPLogger,
-    private val options: Map<String, String>,
+    private val options: Options,
 ) : SymbolProcessor {
     private var invoked = false
 
-    private val packageNameProp = "packageName"
+    private val beeFeatures = ServiceLoader.load(
+        BeeGenerativeFeature::class.java,
+        DataProcessor::class.java.classLoader
+    ).sortedBy { feature -> feature.order() }
 
-    data class DgsDto(
-        val name: String,
-        val properties: List<PropertyDetails>
-    )
-
-    data class PropertyDetails(
-        val name: String,
-        val nonCollectionType: String,
-        val isCollection: Boolean
-    )
-
-    @OptIn(KspExperimental::class)
     override fun process(resolver: Resolver): List<KSAnnotated> {
-        if (invoked) return emptyList()
-
-        val serviceLoader = ServiceLoader.load(
-            BeeGenerativeFeature::class.java,
-            DataProcessor::class.java.classLoader
-        )
-        // if (!options.containsKey(packageNameProp))
-        //     throw Exception("Please provide package name to ksp via `$packageNameProp` property")
-        // val packageName = options.getValue(packageNameProp)
-        //
-        // val foundDtos = mutableListOf<KSClassDeclaration>()
-        //
-        // val dl = resolver
-        //     .getSymbolsWithAnnotation(AutoFetcher::class.qualifiedName!!)
-        //     .mapNotNull { if (it is KSClassDeclaration) it else null }
-        //     .toList()
-        // val dlNames = dl.map { it.simpleName.asString() }
-        // logger.warn("Found DL $dlNames")
-        //
-        // resolver.getAllFiles().forEach { ksFile ->
-        //     ksFile.declarations.forEach { ksDeclaration ->
-        //         if (ksDeclaration is KSClassDeclaration) {
-        //             val ksPackageName = ksDeclaration.packageName.asString()
-        //             if (ksPackageName == packageName)
-        //                 foundDtos.add(ksDeclaration)
-        //         }
-        //
-        //     }
-        // }
-        //
-        // val names = foundDtos.map { it.simpleName.asString() }
-        // logger.warn("Found DTOs $names")
-        // val dtos = transform(foundDtos)
-        // logger.warn("$dtos")
-        //
-        // val dependencies = Dependencies(false, *resolver.getAllFiles().toList().toTypedArray())
-        // for (dataLoader in dl) {
-        //     dataLoader.accept(AutoFetcherVisitor(codeGenerator, dtos, dependencies, logger), Unit)
-        // }
-        // TODO: Modularize approach, currently hardcoded
-        logger.warn("$options")
-        val dl = resolver
-            .getSymbolsWithAnnotation("com.beeproduced.bee.fetched.annotations.BeeFetched")
-            .mapNotNull { if (it is KSClassDeclaration) it else null }
-            .toList()
-        val dlNames = dl.map { it.simpleName.asString() }
-        logger.warn("Found DL $dlNames")
-
-        val packageName = options.getValue("fetchedScanPackage")
-        // val foundDtos = mutableListOf<KSClassDeclaration>()
-        // resolver.getAllFiles().forEach { ksFile ->
-        //     ksFile.declarations.forEach { ksDeclaration ->
-        //         if (ksDeclaration is KSClassDeclaration) {
-        //             val ksPackageName = ksDeclaration.packageName.asString()
-        //             if (ksPackageName == packageName)
-        //                 foundDtos.add(ksDeclaration)
-        //         }
-        //
-        //     }
-        // }
-        val foundDtos = resolver.getDeclarationsFromPackage(packageName)
-            .mapNotNull { if (it is KSClassDeclaration) it else null }
-            .toList()
-
-        val names = foundDtos.map { it.simpleName.asString() }
-        logger.warn("Found DTOs $names")
-        val dtos = transform(foundDtos)
-        // logger.warn("$dtos")
-
-
-        for (feature in serviceLoader) {
-            logger.warn("New feature in ksp: ${feature::class.java.name}")
-            // feature.process()
-            logger.warn(feature::class.java.name)
-            val config = feature.setup(options)
-            logger.warn("Config: $config")
-            val dependencies = Dependencies(false, *resolver.getAllFiles().toList().toTypedArray())
-
-            val packages = mutableMapOf(packageName to foundDtos.toSet())
-
-            val input = BeeGenerativeInput(codeGenerator, dependencies, logger, packages, options)
-            val visitor = feature.visitor(input)
-
-            for (dataLoader in dl) {
-                dataLoader.accept(visitor, Unit)
-            }
-
+        if (beeFeatures.isEmpty()) {
+            logger.warn("No bee.features found")
+            logger.warn("Possible misconfiguration detected")
+            return emptyList()
         }
-        logger.warn("Service loader: ${serviceLoader.count()}")
 
+        // Not every feature supports ksp multiple round processing
+        val features = if (!invoked) {
+            invoked = true
+            beeFeatures
+        }
+        else beeFeatures.filter { it.multipleRoundProcessing() }
 
-        invoked = true
+        // Acquire config per feature & merge them
+        logger.warn("Received Options $options")
+        val config = features
+            .map { it.setup(options) }
+            .let(BeeGenerativeConfig.Companion::merge)
+        logger.warn("Received Config to parse $config")
+
+        // Get symbols mandated by config
+        val symbols = getSymbols(resolver, config)
+
+        // Process feature implementations
+        val dependencies = Dependencies(false, *resolver.getAllFiles().toList().toTypedArray())
+        val input = BeeGenerativeInput(
+            codeGenerator = codeGenerator,
+            dependencies = dependencies,
+            logger = logger,
+            symbols = symbols,
+            options = options
+        )
+        features.forEach { feature ->
+            logger.warn("Processing ${feature::class.java.name} ...")
+            feature.process(input)
+        }
+
+        logger.warn("Process finished")
         return emptyList()
     }
 
+    @OptIn(KspExperimental::class)
+    private fun getSymbols(resolver: Resolver, config: BeeGenerativeConfig): BeeGenerativeSymbols {
+        val packages = mutableMapOf<String, Set<KSClassDeclaration>>()
+        for (p in config.packages) {
+            val symbols = resolver
+                .getDeclarationsFromPackage(p)
+                .mapNotNull { if (it is KSClassDeclaration) it else null }
+                .toSet()
+            packages[p] = symbols
+        }
+        val annotatedBy = mutableMapOf<String, Set<KSClassDeclaration>>()
+        for (ab in config.annotatedBy) {
+            val symbols = resolver
+                .getSymbolsWithAnnotation(ab)
+                .mapNotNull { if (it is KSClassDeclaration) it else null }
+                .toSet()
+            annotatedBy[ab] = symbols
+        }
 
-
-    private fun transform(classes: List<KSClassDeclaration>): Map<String, DgsDto> {
-        return classes.map { ksClass ->
-            val properties = ksClass.getAllProperties().map { property ->
-                val type = property.type.resolve()
-                if (type.declaration.qualifiedName?.asString() == "kotlin.collections.List" && type.arguments.size == 1) {
-                    PropertyDetails(
-                        name = property.simpleName.asString(),
-                        nonCollectionType = type.arguments.first().type?.resolve()?.declaration?.qualifiedName?.asString() ?: "Unknown",
-                        isCollection = true
-                    )
-                } else {
-                    PropertyDetails(
-                        name = property.simpleName.asString(),
-                        nonCollectionType = type.declaration.qualifiedName?.asString() ?: "Unknown",
-                        isCollection = false
-                    )
-                }
-            }.toList()
-            DgsDto(
-                name = ksClass.qualifiedName?.asString() ?: "Unknown",
-                properties = properties
-            )
-        }.associateBy { it.name }
+        return BeeGenerativeSymbols(packages, annotatedBy)
     }
-
 }
